requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.11
ldap3==2.9.1
pyotp==2.9.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0